Move spinner into shared-ui
