+++ bikespace_frontend/src/components/shared-ui/spinner/index.ts
@@ -1,0 +1,1 @@
+export {Spinner} from './Spinner';
