+++ bikespace_frontend/src/components/shared-ui/spinner/spinner.module.scss
